requirements.txt
pyswisseph>=2.10
timezonefinder>=6.0
